moved faction rulesets into individual folders when changing subfactions the combat groups are now cleared
