+++ lib/models/rules/black_talons/black_talons.dart
@@ -1,14 +1,13 @@
 import 'package:gearforce/models/combatGroups/combat_group.dart';
 import 'package:gearforce/models/factions/faction_type.dart';
 import 'package:gearforce/models/mods/factionUpgrades/faction_mod.dart';
-import 'package:gearforce/models/roster/roster.dart';
 import 'package:gearforce/models/rules/rule_set.dart';
 import 'package:gearforce/models/rules/special_unit_filter.dart';
 import 'package:gearforce/models/unit/role.dart';
 import 'package:gearforce/models/unit/unit.dart';
 
-class South extends RuleSet {
-  const South(super.data);
+class BlackTalons extends RuleSet {
+  const BlackTalons(super.data);
 
   @override
   List<Unit> availableUnits({
@@ -19,7 +18,7 @@ class South extends RuleSet {
     return data
         .getUnits(
           baseFactionFilters: [
-            FactionType.South,
+            FactionType.BlackTalon,
             FactionType.Airstrike,
             FactionType.Universal,
             FactionType.Universal_TerraNova,
@@ -33,8 +32,18 @@ class South extends RuleSet {
   }
 
   @override
+  List<SpecialUnitFilter> availableSpecialFilters() {
+    return [
+      const SpecialUnitFilter(
+        text: 'None',
+        filters: [],
+      )
+    ];
+  }
+
+  @override
   List<FactionModification> availableFactionMods(
-      UnitRoster ur, CombatGroup cg, Unit u) {
+      UnitRoster, CombatGroup cg, Unit u) {
     return [];
   }
 }
