+++ lib/models/rules/peace_river/pps.dart
@@ -1,0 +1,30 @@
+import 'package:gearforce/models/combatGroups/combat_group.dart';
+import 'package:gearforce/models/mods/factionUpgrades/faction_mod.dart';
+import 'package:gearforce/models/roster/roster.dart';
+import 'package:gearforce/models/rules/peace_river/peace_river.dart';
+import 'package:gearforce/models/unit/unit.dart';
+
+/*
+PPS - Paxton Private Securities
+The Paxton Private Securities LLC offers private contractors at a good rate. After all, if
+you can’t afford your own army made with the best of Paxton’s offerings, maybe you
+can rent their forces at competitive rates instead. While held in reserve for the highest
+bidder, discounts are available during times of peace to ensure they stay well practiced.
+Z Ex-PRDF: Choose any one upgrade option from the PRDF.
+Z Ex-POC: Choose any one upgrade option from the POC.
+Z Badland’s Soup: One combat group may purchase the following veteran upgrades
+for their models without being veterans; Improved Gunnery, Dual Guns, Brawler,
+Veteran Melee upgrade, or ECCM.
+Z Sub-Contractors: One combat group may be made with models from North,
+South, Peace River, and NuCoal (may include a mix from all four factions) that
+have an armor of 8 or lower.
+*/
+class PPS extends PeaceRiver {
+  PPS(super.data);
+
+  @override
+  List<FactionModification> availableFactionMods(
+      UnitRoster ur, CombatGroup cg, Unit u) {
+    return super.availableFactionMods(ur, cg, u);
+  }
+}
