+++ lib/models/rules/utopia/utopia.dart
@@ -7,8 +7,8 @@ import 'package:gearforce/models/rules/special_unit_filter.dart';
 import 'package:gearforce/models/unit/role.dart';
 import 'package:gearforce/models/unit/unit.dart';
 
-class Caprice extends RuleSet {
-  const Caprice(super.data);
+class Utopia extends RuleSet {
+  const Utopia(super.data);
 
   @override
   List<Unit> availableUnits({
@@ -19,7 +19,7 @@ class Caprice extends RuleSet {
     return data
         .getUnits(
           baseFactionFilters: [
-            FactionType.Caprice,
+            FactionType.Utopia,
             FactionType.Airstrike,
             FactionType.Universal,
             FactionType.Universal_Non_TerraNova,
@@ -30,6 +30,16 @@ class Caprice extends RuleSet {
         )
         .map((uc) => Unit(core: uc))
         .toList();
+  }
+
+  @override
+  List<SpecialUnitFilter> availableSpecialFilters() {
+    return [
+      const SpecialUnitFilter(
+        text: 'None',
+        filters: [],
+      )
+    ];
   }
 
   @override
